requirements.txt
pandas
streamlit
google-auth
gspread
streamlit_echarts
streamlit_autorefresh
